requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
python-socketio==5.11.0
python-engineio==4.9.0
APScheduler==3.10.4
WTForms==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
bcrypt==4.1.2
SQLAlchemy==2.0.35
email-validator==2.1.0
ffmpeg-python==0.2.0

# Optional development dependencies
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.12.0
# flake8==6.1.0
# isort==5.13.0
